# artifacts/api-server/src/routes/auth.ts
import { Router, type IRouter } from "express";
import { eq } from "drizzle-orm";
import { db, usersTable, sessionsTable } from "@workspace/db";
import { SignUpBody, SignInBody, AuthUserSchema } from "@workspace/api-zod";
import {
  hashPassword,
  verifyPassword,
  generateSessionToken,
  hashSessionToken,
  SESSION_COOKIE_NAME,
  SESSION_TTL_MS,
} from "../lib/auth";
import { requireAuth } from "../middlewares/require-auth";

const router: IRouter = Router();

const COOKIE_OPTIONS = {
  httpOnly: true,
  sameSite: "lax" as const,
  secure: process.env.NODE_ENV === "production",
  path: "/",
};

async function createSessionAndRespond(
  res: import("express").Response,
  user: { id: string; name: string; email: string; academicYear: string | null },
  statusCode: number,
): Promise<void> {
  const token = generateSessionToken();
  await db.insert(sessionsTable).values({
    id: hashSessionToken(token),
    userId: user.id,
    expiresAt: new Date(Date.now() + SESSION_TTL_MS),
  });

  res.cookie(SESSION_COOKIE_NAME, token, {
    ...COOKIE_OPTIONS,
    maxAge: SESSION_TTL_MS,
  });

  const body = AuthUserSchema.parse({
    id: user.id,
    name: user.name,
    email: user.email,
    academicYear: user.academicYear,
  });
  res.status(statusCode).json(body);
}

router.post("/auth/signup", async (req, res) => {
  const parsed = SignUpBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ message: parsed.error.issues[0]?.message ?? "Invalid request" });
    return;
  }
  const { name, email, password, academicYear } = parsed.data;

  const existing = await db
    .select({ id: usersTable.id })
    .from(usersTable)
    .where(eq(usersTable.email, email))
    .limit(1);
  if (existing.length > 0) {
    res.status(409).json({ message: "That email is already registered" });
    return;
  }

  const passwordHash = await hashPassword(password);
  const [user] = await db
    .insert(usersTable)
    .values({ name, email, passwordHash, academicYear })
    .returning();

  if (!user) {
    res.status(500).json({ message: "Failed to create account" });
    return;
  }

  await createSessionAndRespond(res, user, 201);
});

router.post("/auth/signin", async (req, res) => {
  const parsed = SignInBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ message: "Invalid request" });
    return;
  }
  const { email, password } = parsed.data;

  const [user] = await db
    .select()
    .from(usersTable)
    .where(eq(usersTable.email, email))
    .limit(1);

  if (!user || !(await verifyPassword(password, user.passwordHash))) {
    res.status(401).json({ message: "Invalid email or password" });
    return;
  }

  await createSessionAndRespond(res, user, 200);
});

router.post("/auth/logout", async (req, res) => {
  const token = req.cookies?.[SESSION_COOKIE_NAME] as string | undefined;
  if (token) {
    await db
      .delete(sessionsTable)
      .where(eq(sessionsTable.id, hashSessionToken(token)));
  }
  res.clearCookie(SESSION_COOKIE_NAME, COOKIE_OPTIONS);
  res.status(204).end();
});

router.get("/auth/me", requireAuth, (req, res) => {
  const body = AuthUserSchema.parse(req.user);
  res.json(body);
});

export default router;