# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Provisional — regenerate with `pnpm --filter @workspace/api-spec run codegen`
 * once dependencies are installed. Matches openapi.yaml auth schemas.
 */

export const AuthUserSchema = zod.object({
  id: zod.string().uuid(),
  name: zod.string(),
  email: zod.string().email(),
  academicYear: zod.string().nullish(),
});

export const ErrorResponseSchema = zod.object({
  message: zod.string(),
});

export const SignUpBody = zod.object({
  name: zod.string().min(1),
  email: zod.string().email(),
  password: zod.string().min(8),
  academicYear: zod.string().optional(),
});

export const SignInBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const UpdateProgressBody = zod.object({
  progressPercent: zod.number().int().min(0).max(100),
});

export const AddStudyTimeBody = zod.object({
  seconds: zod.number().int().min(1),
});

export const IncrementGoalBody = zod.object({
  type: zod.enum(["chapters", "mcqs", "videos"]),
});