# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Provisional — regenerate with `pnpm --filter @workspace/api-spec run codegen`
 * once dependencies are installed. Matches openapi.yaml `ErrorResponse`.
 */

export interface ErrorResponse {
  message: string;
}