# lib/api-zod/src/generated/types/signInRequest.ts
/**
 * Provisional — regenerate with `pnpm --filter @workspace/api-spec run codegen`
 * once dependencies are installed. Matches openapi.yaml `SignInRequest`.
 */

export interface SignInRequest {
  email: string;
  password: string;
}